Remove stuck status when element is scrolled out of view
If we have a container with a sticky element, and we scroll past the bottom edge of that container, then the element will still be stuck at the bottom of its container. However, we might want to remove some styles in such a case. For example, we may want to remove a shadow from a header scrolled out of view; otherwise, it might still be visible at the top of the screen for a split second even though the header it belongs to is gone.